feat: react on children which has been inserted after the relation
Given a parent receives a foreign key which can not be resolved immediately as the child has not been inserted yet, that child will never be handled at all. Since we can track added events on the child we can use the id of those dangling children to start the add procedure once the document has been inserted.
Generally this should never happen if the foreign key relationship consistency e.g. insertion order is kept. For those cases where it wasn't this should solve the issue.