fix: diff changes from change stream with op update/replace with the currently published meteor document
As the updateDescription can not be used to send the diff due to implementation incompatibility and the full document might not represent the state of the updateDescription we now rely on the current published document to compute the changeset. This also solves the issue with replaced documents which weren't properly supported yet.